Step 2: trivial server-side rendering with Express
RUN: `npm run start` and go to `http://localhost:3000`.

- A little refactoring: move Hello component to separate file etc
- Render the same Hello component server-side with ReactDOMServer

Now we are rendering the same Hello component both on client and server:
with express server we pre-render the Hello component on the server, and
server the client with rendered html, and with webpack we continue to
bundle client.js into ES5 code that the browser would understand,
just as we did at previous step.