Step 3a: switch to CSS modules
Everybody loves CSS modules, and the great news is that they come free with Webpack.
The bad news is that we can't use them with server-side rendering, as we don't use
Webpack when rendering on the server-side.
So at this step we broke everything, and the only way to continue from here, is to
start using Webpack to pre-build code for server-side rendering too, and that's
what we'll do at the next step.