Step 4a: asyncronously fetching data
Now let's fetch some data asyncronously. We'll use isomorphic-fetch,
as it works both on client and server our of the box.

Fetching data works just fine on the server, but the problem is that
on the server we didn't wait for fetch to finish fetching the data
before sending the response, so our pre-rendered html doesn't have any
async data when it arrives.
Let's try to fix it in the next step.